Move celery dependent parts out into their own module and make default API routing information more easily accessible.
